requirements.txt
altair==5.2.0
attrs==23.1.0
audioread==3.0.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
duckduckgo-images-api==1.0.0
ffmpeg==1.4
ffmpeg-python==0.2.0
fonttools==4.48.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.40
idna==3.6
importlib-metadata==6.11.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
msgpack==1.0.7
numba==0.59.0
numpy==1.26.4
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
platformdirs==4.2.0
pooch==1.8.0
protobuf==4.25.1
pyarrow==14.0.2
PyAudio==0.2.14
pycparser==2.21
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.32.0
requests==2.31.0
resampy==0.4.2
rich==13.7.0
rpds-py==0.15.2
scikit-learn==1.4.0
scipy==1.12.0
setuptools==69.0.3
six==1.16.0
smmap==5.0.1
sounddevice==0.4.6
soundfile==0.12.1
soxr==0.3.7
streamlit==1.29.0
tenacity==8.2.3
threadpoolctl==3.2.0
tinydb==4.8.0
tinydb-serialization==2.1.1
tinytag==1.10.1
toml==0.10.2
toolz==0.12.0
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
